Refactor calendar code
