--- src/components/elements/calendar/events/EventCreator.tsx
@@ -1,7 +1,0 @@
-export const EventCreator = () => {
-  return (
-    <div>
-      <p>empty</p>
-    </div>
-  );
-};
