+++ src/components/elements/calendar/events/index.ts
@@ -2,4 +2,3 @@ export * from "./Event";
 export * from "./AllDayEvent";
 export * from "./normal";
 export * from "./MultiDayEvent";
-export * from "./EventCreator";
