+++ src/components/elements/calendar/index.ts
@@ -3,3 +3,4 @@ export * from "./CalendarDate";
 export * from "./events";
 export * from "./CalendarFilterButtons";
 export * from "./calendarUtils";
+export * from "./event-creation";
