+++ src/components/elements/index.ts
@@ -5,5 +5,4 @@ export * from "./button";
 export * from "./Modal";
 export * from "./ResizablePanel";
 export * from "./user";
-export * from "./time-date-pickers";
 export * from "./moving-box";
